+++ src/pixelator/pna/analysis/segmentation/__init__.py
@@ -6,5 +6,11 @@ Copyright © 2026 Pixelgen Technologies AB.
 from pixelator.pna.analysis.segmentation.distance import distance_from_node_set
 from pixelator.pna.analysis.segmentation.partition import partition_counts
 from pixelator.pna.analysis.segmentation.protein_weights import cc_protein_weights
+from pixelator.pna.analysis.segmentation.segment import segment_cell
 
-__all__ = ["cc_protein_weights", "distance_from_node_set", "partition_counts"]
+__all__ = [
+    "cc_protein_weights",
+    "distance_from_node_set",
+    "partition_counts",
+    "segment_cell",
+]